Fix tests and standardise openapi return values
